Use PI definition from math.h

+++ trunk/lib/wormux/include/WORMUX_point.h
@@ -28,7 +28,7 @@
 #include <WORMUX_vector2.h>
 
 typedef Vector2<int32_t> Point2i;
-typedef Vector2<Double>  Point2f;
+typedef Vector2<float>   Point2f;
 typedef Vector2<Double>  Point2d;
 
 #define POINT2D_2_POINT2I(point) \
